Separar classe de Message e organizar o array de mensagens
